+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/data/local/item/ItemRoomRepository.kt
@@ -21,7 +21,7 @@ class ItemRoomRepository
                 }
             }
 
-        override fun getItemStream(id: Int): Flow<ItemModel> = itemDao.getItemById(id).map { it.asModel() }
+        override fun getItemStream(id: Long): Flow<ItemModel> = itemDao.getItemById(id).map { it.asModel() }
 
         override suspend fun insertItem(item: ItemModel) = itemDao.insert(item.asEntity())
